Title: 앞의 두 수를 더한 수열 2, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240322/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    int table[91];
+    long long table[91];
     int n;
     cin>>n;
     table[1]=1;
